Remove leading newlines from docstrings
If pybind11-stubgen is run on a normal python docstring it will convert this
```py
"""
This is a docstring.
"""
```
To this
```py
"""

This is a docstring.
"""
```

The leading newline is interpreted as part of the docstring and pybind11-stubgen will add another newline as part of its formatting.
Stripping these leading newlines should have no effect in docstrings defined using pybind11 unless the user adds a leading newline for some reason.